Add ability to invoke arbitrary Salt command via a method effector.
The command is executed as a task (activity).

If the command succeeds the effector will return the standard output of the command.

If the command fails, an exception will be thrown (resulting in 500 error if invoked through REST API), and the standard error of the command will be available for inspection via the activity history of the entity.